refactor(ui): extract IssueDetailPage sidebar panels and move modal
